refactor(workspace-strategy): extract strategy/realization types from @paperclipai/shared
