+++ packages/workspace-strategy/src/index.ts
@@ -1,1 +1,8 @@
-export const PACKAGE_NAME = "@paperclipai/workspace-strategy";
+export {
+  type ExecutionWorkspaceStrategyType,
+  type ExecutionWorkspaceStrategy,
+  type WorkspaceRealizationTransport,
+  type WorkspaceRealizationSyncStrategy,
+  type WorkspaceRealizationRequest,
+  type WorkspaceRealizationRecord,
+} from "./types.js";
